Safari-611.1.10/LayoutTests/loader/navigation-policy/should-open-external-urls/user-gesture-window-open-without-flag.html

<script>
if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.dumpPolicyDelegateCallbacks();
    testRunner.queueLoad("resources/user-gesture-window-open-to-notify-done.html");
}
</script>
This page opens a 2nd page with the "should open external URLs" flag set to false.<br>
That page has a link that - when clicked - does a window.open to open a 3rd page.<br>
That 3rd page should have the flag set to true because it was from a user gesture.<br>
