Safari-611.1.10/LayoutTests/loader/navigation-policy/should-open-external-urls/user-gesture-window-open-without-flag-expected.txt

 - decidePolicyForNavigationAction
<NSURLRequest URL resources/user-gesture-window-open-to-notify-done.html, main document URL resources/user-gesture-window-open-to-notify-done.html, http method GET> is main frame - yes should open URLs externally - no
 - decidePolicyForNavigationAction
<NSURLRequest URL notify-done.html, main document URL notify-done.html, http method GET> is main frame - yes should open URLs externally - yes
Click to do a window.open
